@@ -5,20 +5,23 @@ version = "0.10.5"
 enable_black = true
 enable_flake8 = true
 enable_isort = true
-enable_mypy = false    # TODO: eliminate errors and enable at CI
-mypy_preset = "entry"  # TODO: "strict"
+enable_mypy = true
+mypy_preset = "entry" # TODO: "strict"
+mypy_plugins = [
+  { function = "numpy.typing.mypy_plugin" },
+  { function = "pydantic.mypy" },
+]
 line_length = 88
 py_version = "py311"
 isort_known_first_party = ["voicevox_engine"]
 isort_known_third_party = ["numpy"]
 [[tool.pysen.lint.mypy_targets]]
-  paths = [".", "voicevox_engine/"]
+paths = ["run.py", "voicevox_engine/"]
 
 [tool.black] # automatically generated by pysen
 # pysen ignores and overwrites any modifications
 line-length = 88
-target-version = ["py310", "py311"]
-
+target-version = ["py311"]
 
 [tool.isort] # automatically generated by pysen
 # pysen ignores and overwrites any modifications
