@@ -1,4 +1,4 @@
-from typing import Sequence
+from collections.abc import Collection
 
 from semver.version import Version
 
@@ -7,7 +7,7 @@ def parse_core_version(version: str) -> Version:
     return Version.parse(version)
 
 
-def get_latest_core_version(versions: Sequence[str]) -> str:
+def get_latest_core_version(versions: Collection[str]) -> str:
     if len(versions) == 0:
         raise Exception("versions must be non-empty.")
 
