@@ -2,7 +2,12 @@ import json
 from pathlib import Path
 from typing import TYPE_CHECKING, Dict, List, Tuple
 
-from voicevox_engine.metas.Metas import CoreSpeaker, EngineSpeaker, Speaker, StyleInfo
+from voicevox_engine.metas.Metas import (
+    CoreSpeaker,
+    EngineSpeaker,
+    Speaker,
+    SpeakerStyle,
+)
 
 if TYPE_CHECKING:
     from voicevox_engine.core_adapter import CoreAdapter
@@ -54,7 +59,9 @@ class MetasStore:
         ]
 
 
-def construct_lookup(speakers: List[Speaker]) -> Dict[int, Tuple[Speaker, StyleInfo]]:
+def construct_lookup(
+    speakers: List[Speaker],
+) -> Dict[int, Tuple[Speaker, SpeakerStyle]]:
     """
     スタイルID に話者メタ情報・スタイルメタ情報を紐付ける対応表を生成
     Parameters
@@ -63,10 +70,10 @@ def construct_lookup(speakers: List[Speaker]) -> Dict[int, Tuple[Speaker, StyleI
         話者メタ情報
     Returns
     -------
-    ret : Dict[int, Tuple[Speaker, StyleInfo]]
+    ret : Dict[int, Tuple[Speaker, SpeakerStyle]]
         スタイルID に話者メタ情報・スタイルメタ情報が紐付いた対応表
     """
-    lookup_table = dict()
+    lookup_table: dict[int, tuple[Speaker, SpeakerStyle]] = dict()
     for speaker in speakers:
         for style in speaker.styles:
             lookup_table[style.id] = (speaker, style)
