@@ -1,6 +1,6 @@
 from pathlib import Path
 
-import yaml
+import yaml  # type:ignore[import-untyped]
 
 from ..utility import get_save_dir
 from .Setting import Setting
@@ -28,12 +28,11 @@ class SettingLoader:
             # 指定された設定ファイルから値を取得
             setting = yaml.safe_load(self.setting_file_path.read_text(encoding="utf-8"))
 
-        setting = Setting(
-            cors_policy_mode=setting["cors_policy_mode"],
+        # FIXME: 型チェックと例外処理を追加する
+        return Setting(
+            cors_policy_mode=setting["cors_policy_mode"],  # type: ignore[arg-type]
             allow_origin=setting["allow_origin"],
         )
-
-        return setting
 
     def dump_setting_file(self, settings: Setting) -> None:
         settings_dict = settings.dict()
