@@ -9,7 +9,6 @@ from uuid import UUID, uuid4
 import numpy as np
 import pyopenjtalk
 from fastapi import HTTPException
-from pydantic import conint
 
 from .model import UserDictWord, WordTypes
 from .part_of_speech_data import MAX_PRIORITY, MIN_PRIORITY, part_of_speech_data
@@ -436,16 +435,18 @@ def _search_cost_candidates(context_id: int) -> List[int]:
     raise HTTPException(status_code=422, detail="品詞IDが不正です")
 
 
-def _cost2priority(context_id: int, cost: conint(ge=-32768, le=32767)) -> int:
+def _cost2priority(context_id: int, cost: int) -> int:
+    assert -32768 <= cost <= 32767
     cost_candidates = _search_cost_candidates(context_id)
     # cost_candidatesの中にある値で最も近い値を元にpriorityを返す
     # 参考: https://qiita.com/Krypf/items/2eada91c37161d17621d
     # この関数とpriority2cost関数によって、辞書ファイルのcostを操作しても最も近いpriorityのcostに上書きされる
-    return MAX_PRIORITY - np.argmin(np.abs(np.array(cost_candidates) - cost))
+    return MAX_PRIORITY - np.argmin(
+        np.abs(np.array(cost_candidates) - cost)
+    )  # type:ignore [return-value]
 
 
-def _priority2cost(
-    context_id: int, priority: conint(ge=MIN_PRIORITY, le=MAX_PRIORITY)
-) -> int:
+def _priority2cost(context_id: int, priority: int) -> int:
+    assert MIN_PRIORITY <= priority <= MAX_PRIORITY
     cost_candidates = _search_cost_candidates(context_id)
     return cost_candidates[MAX_PRIORITY - priority]
